+++ CHANGELOG.md
@@ -97,6 +97,7 @@
 - PR #6780 Move `cudf::cast` tests to separate test file
 - PR #6789 Rename `unary_op` to `unary_operator`
 - PR #6770 Support building decimal columns with Table.TestBuilder
+- PR #6805 Enable copy_if for fixed-point decimal columns
 
 ## Bug Fixes